refactor(peer-id): split crate into focused internal modules
