+++ packages/peer-id/src/lib.rs
@@ -2,6 +2,8 @@
 
 #![allow(clippy::module_name_repetitions)]
 
+mod peer_client;
 mod peer_id;
 
-pub use self::peer_id::{PeerClient, PeerId};
+pub use self::peer_client::PeerClient;
+pub use self::peer_id::PeerId;
